requirements.txt
attrs==19.3.0
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.0
hydra-joblib-launcher==1.1.5
hydra-submitit-launcher==1.2.0
git+https://github.com/captain-pool/hydra-wandb-sweeper.git
jax==0.2.11
jaxlib==0.1.63
lightgbm==4.0.0
matplotlib==3.4.2
numba==0.56.4
numpy==1.19.4
pandas==1.1.5
scikit-learn==1.1.3
seaborn==0.12.2
shap==0.41.0
tensorflow==2.5.0
tensorflow_privacy==0.5.2
tensorflow_federated==0.19.0
wandb==0.12.21